+++ src/thread/creation/Program2.java
@@ -1,4 +1,4 @@
-package programs;
+package thread.creation;
 
 
 /*
